requirements.txt
streamlit
gTTS
openai
coqui-tts
pydub
soundfile
matplotlib
numpy
pytest
python-dotenv

